+++ src/components/NavBar/index.tsx
@@ -1,106 +1,66 @@
-import {Link} from "react-router-dom";
-import {
-    PAGINATION_ROUTE,
-    FORM_ROUTE,
-    BULBASAUR_ROUTE,
-    IVYSAUR_ROUTE, POKEMONS_ROUTE,
-} from "../../app/routes/config";
-import type {MenuProps} from "antd";
-import {Menu, Button, Switch} from "antd";
-import {useEffect, useState} from "react";
-import styled from "styled-components";
+import type { MenuProps } from "antd";
+import { Button, Switch } from "antd";
+import { useEffect, useState } from "react";
+import { StyledMenu } from "./style.tsx";
+import { authItems, defaultItems } from "./items.tsx";
 
 interface NavBarProps {
-    isAuth: boolean;
-    setIsAuth: (isAuthenticated: boolean) => void;
+  isAuth: boolean;
+  setIsAuth: (isAuthenticated: boolean) => void;
 }
 
 type ThemeType = "dark" | "light";
 
-export const StyledMenu = styled(Menu)`
-    background: inherit;
-    color: inherit;
-    margin-bottom: 0.5rem;
+const NavBar: React.FC<NavBarProps> = ({ isAuth, setIsAuth }) => {
+  const [currentTheme, setCurrentTheme] = useState<ThemeType>(
+    window.matchMedia("(prefers-color-scheme: light)").matches
+      ? "light"
+      : "dark",
+  );
 
-    .ant-menu-item-active {
-        color: inherit !important;
-    }
-`;
+  const changeColorTheme = () => {
+    const newTheme = currentTheme === "light" ? "dark" : "light";
+    setCurrentTheme(newTheme);
+  };
 
-const NavBar: React.FC<NavBarProps> = ({isAuth, setIsAuth}) => {
-    const [currentTheme, setCurrentTheme] = useState<ThemeType>(
-        window.matchMedia("(prefers-color-scheme: light)").matches
-            ? "light"
-            : "dark"
-    );
+  const setInitialTheme = (theme: ThemeType) => {
+    document.documentElement.setAttribute("data-theme", theme);
+  };
 
-    const changeColorTheme = () => {
-        const newTheme = currentTheme === "light" ? "dark" : "light";
-        setCurrentTheme(newTheme);
-    };
+  const toggleAuth = () => {
+    setIsAuth(!isAuth);
+  };
 
-    const setInitialTheme = (theme: ThemeType) => {
-        document.documentElement.setAttribute("data-theme", theme);
-    };
+  const authButtonText = isAuth ? "Log out" : "Log in";
 
-    const toggleAuth = () => {
-        setIsAuth(!isAuth);
-    };
+  const actionItems: MenuProps["items"] = [
+    {
+      label: (
+        <Switch
+          onChange={changeColorTheme}
+          checkedChildren="Dark"
+          unCheckedChildren="Light"
+          defaultChecked={currentTheme === "dark"}
+        />
+      ),
+      key: "theme",
+    },
+    {
+      label: <Button onClick={toggleAuth}>{authButtonText}</Button>,
+      key: "auth",
+    },
+  ];
 
-    const authButtonText = isAuth ? "Log out" : "Log in";
+  const items: MenuProps["items"] = [];
+  items.push(...(defaultItems as []));
+  if (isAuth) items.push(...(authItems as []));
+  items.push(...actionItems);
 
-    const items: MenuProps["items"] = [
-        {
-            label: <Link to={PAGINATION_ROUTE}>Pagination</Link>,
-            key: "pagination",
-        },
-        {
-            label: <Link to={BULBASAUR_ROUTE}>Bulbasaur</Link>,
-            key: "bulbasaur",
-        },
-        {
-            label: <Link to={IVYSAUR_ROUTE}>Ivysaur</Link>,
-            key: "ivysaur",
-        },
-        {
-            label: <Link to={POKEMONS_ROUTE}>Pokemons</Link>,
-            key: "pokemons",
-        },
-    ];
+  useEffect(() => {
+    setInitialTheme(currentTheme);
+  }, [currentTheme]);
 
-    const authItems: MenuProps["items"] = [
-        {
-            label: <Link to={FORM_ROUTE}>Form</Link>,
-            key: "Form",
-        },
-    ];
-
-    const actionItems: MenuProps["items"] = [
-        {
-            label: (
-                <Switch
-                    onChange={changeColorTheme}
-                    checkedChildren="Dark"
-                    unCheckedChildren="Light"
-                    defaultChecked={currentTheme === "dark"}
-                />
-            ),
-            key: "theme",
-        },
-        {
-            label: <Button onClick={toggleAuth}>{authButtonText}</Button>,
-            key: "auth",
-        },
-    ];
-
-    if (isAuth) items.push(...authItems);
-    items.push(...actionItems);
-
-    useEffect(() => {
-        setInitialTheme(currentTheme);
-    }, [currentTheme]);
-
-    return <StyledMenu mode="horizontal" items={items}/>;
+  return <StyledMenu mode="horizontal" items={items} />;
 };
 
 export default NavBar;
