+++ src/main.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
@@ -9,5 +8,5 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
   <BrowserRouter>
     <GlobalStyles />
     <App />
-  </BrowserRouter>
+  </BrowserRouter>,
 );
